Restrict get_pcb_component to single exact-refdes lookup
Renamed from get_pcb_components (plural, regex multi-match) to
get_pcb_component (singular, exact string match). Returns a single
ComponentResult directly instead of a wrapper with matches array.
Prevents token overflow from broad queries like "^R\d+" matching
hundreds of components.
